lance-integration/tools/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lance-distributed"
version = "0.1.0"
description = "Python client for Lance-Ballista distributed multimodal retrieval"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.0",
    "pyarrow>=14.0",
    "numpy",
]

[project.optional-dependencies]
dev = ["pytest", "lance>=4.0", "pyyaml"]
